+++ 1.21.3/Common/src/main/java/fuzs/puzzleslib/api/core/v1/ModConstructor.java
@@ -120,4 +120,11 @@ public interface ModConstructor extends BaseModConstructor {
     default void onAddDataPackFinders(final PackRepositorySourcesContext context) {
         // NO-OP
     }
+
+    /**
+     * @param context register data pack-driven dynamic registries
+     */
+    default void onDataPackRegistriesContext(final DataPackRegistriesContext context) {
+        // NO-OP
+    }
 }